[chore] 더미데이터 파일 분리
Fixes: #29 Related to: #15
